+++ src/app/(normal)/article/[id]/pages/DetailPage.tsx
@@ -7,9 +7,9 @@ import { EVMExplorerResolver } from '@masknet/web3-providers';
 import { ChainId } from '@masknet/web3-shared-evm';
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query';
 import DOMPurify from 'dompurify';
-import { compact, first } from 'lodash-es';
+import { compact } from 'lodash-es';
 import { useRouter } from 'next/navigation.js';
-import { useEffect, useRef } from 'react';
+import { useRef } from 'react';
 import urlcat from 'urlcat';
 import { useDarkMode, useDocumentTitle } from 'usehooks-ts';
 import { checksumAddress } from 'viem';
@@ -28,7 +28,6 @@ import { fetchJSON } from '@/helpers/fetchJSON.js';
 import { useComeBack } from '@/hooks/useComeback.js';
 import { PreviewMediaModalRef } from '@/modals/controls.js';
 import { FireflyArticleProvider } from '@/providers/firefly/Article.js';
-import type { ParagraphJSONContent } from '@/providers/paragraph/type.js';
 import { ArticlePlatform } from '@/providers/types/Article.js';
 import type { Attachment } from '@/providers/types/SocialMedia.js';
 import type { ResponseJSON } from '@/types/index.js';
@@ -77,51 +76,6 @@ export function ArticleDetailPage({ params: { id: articleId } }: PageProps) {
     });
 
     useDocumentTitle(article ? createPageTitle(t`Post by ${article.author.handle}`) : SITE_NAME);
-
-    useEffect(() => {
-        if (!ref.current) return;
-        const paragraphObserver = new MutationObserver(() => {
-            if (!article?.json) return;
-            const json = JSON.parse(article.json) as { content: ParagraphJSONContent[] };
-
-            const twitterEmbeds = json.content.filter((x) => x.type === 'twitter');
-
-            twitterEmbeds.forEach((x) => {
-                const poster = x.attrs?.tweetData?.video?.poster;
-                const videoSrc = first(
-                    x.attrs?.tweetData?.video?.variants.filter((video) => video.type === 'video/mp4'),
-                )?.src;
-                const img = ref.current?.querySelector(`img[src="${poster}"]`);
-
-                if (!img || !videoSrc) return;
-
-                const videoNode = document.createElement('video');
-                videoNode.src = videoSrc;
-                videoNode.controls = true;
-                img?.replaceWith(videoNode);
-            });
-
-            const svgEmbeds = json.content.filter((x) => x.type === 'figure');
-
-            svgEmbeds.forEach((x) => {
-                x.content.forEach((svg) => {
-                    if (svg.attrs?.nextheight && svg.attrs?.nextwidth) {
-                        const node = ref.current?.querySelector(`img[src="${svg.attrs.src}"]`);
-                        node?.setAttribute('height', `${svg.attrs.nextheight}px`);
-                        node?.setAttribute('width', `${svg.attrs.nextwidth}px`);
-                    }
-
-                    return;
-                });
-            });
-        });
-
-        paragraphObserver.observe(ref.current, { subtree: true, childList: true });
-
-        return () => {
-            paragraphObserver.disconnect();
-        };
-    }, [article?.json, article?.content]);
 
     if (!article) return null;
